Fixed errors and warnings from vagrant

+++ project2/tcp_client.cpp
@@ -110,7 +110,7 @@ bool TCP_Client::receiveFile(){
             else if (seq == m_expected_seq) {
 
                 // Write the data to the file
-			    for(int i = 0; i < data_size; i++){
+			    for(ssize_t i = 0; i < data_size; i++){
 				    outputFile << data->at(i);
 			    }
                 m_expected_seq = (m_expected_seq + m_packet->getLength() - HEADER_SIZE) % MAX_SEQ;
@@ -120,12 +120,12 @@ bool TCP_Client::receiveFile(){
                 bool found = true;
                 while (found){
                     found = false;
-                    for (int i = 0; i < packet_buffer.size(); i++) {
+                    for (ssize_t i = 0; i < packet_buffer.size(); i++) {
                         if(m_expected_seq == packet_buffer[i]->getHeader().fields[SEQ]) {
                             found = true;
                             m_expected_seq = (m_expected_seq + packet_buffer[i]->getLength() - HEADER_SIZE) % MAX_SEQ;
                             data = packet_buffer[i]->getData();
-                            for(int i = 0; i < data_size; i++){
+                            for(ssize_t i = 0; i < data_size; i++){
                                 outputFile << data->at(i);
                             }
                             delete packet_buffer[i];
